requirements.txt
librosa==0.10.1
matplotlib==3.8.4
numpy==1.24.3
pandas==2.2.2
python-dotenv==1.0.1
scikit_learn==1.4.2
seaborn==0.13.2
spotipy==2.23.0
tqdm==4.65.0
transformers==4.40.1
xgboost==2.0.3
